pyproject.toml
[tool.poetry]
name = "rltrading"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.4.3"
lemon = "^1.1.0"
python-dotenv = "^0.20.0"
pydantic = "^1.9.1"
black = "^22.6.0"
paho-mqtt = "^1.6.1"
plotly = "^5.9.0"
candlestick-chart = "^2.1.0"
dash = "^2.6.1"
plotext = "^5.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
